Robots can now be set to not fire. Robots stay in idle animation if speed is 1 or lower. Now includes level progression by Argus. Other misc fixes.

+++ berzerk/state_options.cpp
@@ -152,7 +152,7 @@ void StateOptions::HandleInput()
 				{
 					joystickInput = true;
 					selectedButton++;
-					if( selectedButton >= buttons.size() )
+					if( (unsigned int) selectedButton >= buttons.size() )
 						selectedButton = buttons.size() - 1;
 				}
 				else
